+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/bloomery/recipe/BloomeryRecipeBuilder.java
@@ -19,6 +19,7 @@ public class BloomeryRecipeBuilder
         this.output,
         this.input,
         this.burnTimeTicks,
+        this.experience,
         this.failureChance,
         this.bloomYieldMin,
         this.bloomYieldMax,